+++ packages/atomic/src/components/commerce/facets/atomic-commerce-category-facet/atomic-commerce-category-facet.tsx
@@ -29,7 +29,6 @@ import {FacetInfo} from '../../../common/facets/facet-common-store';
 import {FacetContainer} from '../../../common/facets/facet-container/facet-container';
 import {FacetGuard} from '../../../common/facets/facet-guard';
 import {FacetHeader} from '../../../common/facets/facet-header/facet-header';
-import {FacetPlaceholder} from '../../../common/facets/facet-placeholder/facet-placeholder';
 import {announceFacetSearchResultsWithAriaLive} from '../../../common/facets/facet-search/facet-search-aria-live';
 import {FacetSearchInput} from '../../../common/facets/facet-search/facet-search-input';
 import {FacetSearchMatches} from '../../../common/facets/facet-search/facet-search-matches';
@@ -65,13 +64,15 @@ export class AtomicCategoryFacet implements InitializableComponent<Bindings> {
    * The category facet controller instance.
    */
   @Prop() public facet!: CategoryFacet;
+  /**
+   * Specifies whether the facet is collapsed.
+   */
+  @Prop({reflect: true, mutable: true}) public isCollapsed = false;
 
   @BindStateToController('facet')
   @State()
   public facetState!: CategoryFacetState;
   @State() public error!: Error;
-
-  @State() private isCollapsed = false;
 
   private resultIndexToFocusOnShowMore = 0;
   private showLessFocus?: FocusTargetController;
@@ -403,7 +404,7 @@ export class AtomicCategoryFacet implements InitializableComponent<Bindings> {
         hasError={hasError}
         hasResults={values.length > 0}
       >
-        {firstSearchExecuted ? (
+        {
           <FacetContainer>
             {this.renderHeader()}
             {!this.isCollapsed && [
@@ -445,9 +446,7 @@ export class AtomicCategoryFacet implements InitializableComponent<Bindings> {
               ),
             ]}
           </FacetContainer>
-        ) : (
-          <FacetPlaceholder isCollapsed={this.isCollapsed} numberOfValues={8} />
-        )}
+        }
       </FacetGuard>
     );
   }