+++ packages/atomic/src/components/search/atomic-sort-dropdown/atomic-sort-dropdown.new.stories.tsx
@@ -20,7 +20,7 @@ export default meta;
 export const Default: Story = {
   name: 'atomic-sort-dropdown',
   args: {
-    default: `
+    'slots-default': `
       <atomic-sort-expression
         label="relevance"
         expression="relevancy"