+++ packages/headless/src/controllers/triggers/headless-query-trigger.ts
@@ -10,41 +10,8 @@ import {updateIgnoreQueryTrigger} from '../../features/triggers/triggers-actions
 import {triggerReducer as triggers} from '../../features/triggers/triggers-slice';
 import {TriggerSection, QuerySection} from '../../state/state-sections';
 import {loadReducerError} from '../../utils/errors';
-import {buildController, Controller} from '../controller/headless-controller';
-
-/**
- * The `QueryTrigger` controller handles query triggers. A [Query trigger](https://docs.coveo.com/en/3413#query) query pipeline rule lets you define a query to be automatically executed against the index when a certain condition is met.
- */
-export interface QueryTrigger extends Controller {
-  /**
-   * The state of the `QueryTrigger` controller.
-   */
-  state: QueryTriggerState;
-  /**
-   * Undoes a query trigger's correction.
-   */
-  undo(): void;
-}
-
-/**
- * A scoped and simplified part of the headless state that is relevant to the `QueryTrigger` controller.
- */
-export interface QueryTriggerState {
-  /**
-   * The new query to perform a search with after receiving a query trigger.
-   */
-  newQuery: string;
-
-  /**
-   * The query used to perform the search that received a query trigger in its response.
-   */
-  originalQuery: string;
-
-  /**
-   * A boolean to specify if the controller was triggered resulting in a modification to the query.
-   */
-  wasQueryModified: boolean;
-}
+import {buildController} from '../controller/headless-controller';
+import {QueryTrigger} from '../core/triggers/headless-core-query-trigger';
 
 /**
  * Creates a `QueryTrigger` controller instance.