+++ packages/headless/src/features/commerce/facets/facet-set/facet-set-slice.ts
@@ -370,7 +370,6 @@ export const commerceFacetSetReducer = createReducer(
           return;
         }
 
-        // TODO: KIT-3226 No need for this function if the values in the payload already contains appropriate parameters
         request.values = convertToDateRangeRequests(values);
         request.numberOfValues = values.length;
       })
@@ -628,9 +627,9 @@ function insertNewValue(
 }
 
 function setAllFacetValuesToIdle(state: CommerceFacetSetState) {
-  Object.values(state).forEach((facet) => {
-    facet.request.values.forEach((value) => (value.state = 'idle'));
-  });
+  Object.values(state).forEach((facet) =>
+    handleDeselectAllFacetValues(facet.request)
+  );
 }
 
 function clearAllFacetValues(state: CommerceFacetSetState) {