+++ packages/headless/src/features/commerce/pagination/pagination-slice.test.ts
@@ -1,6 +1,7 @@
 import {buildSearchResponse} from '../../../test/mock-commerce-search';
-import {buildFetchProductListingV2Response} from '../../../test/mock-product-listing-v2';
+import {buildFetchProductListingResponse} from '../../../test/mock-product-listing';
 import {buildMockRecommendationsResponse} from '../../../test/mock-recommendations';
+import {deselectAllBreadcrumbs} from '../../breadcrumb/breadcrumb-actions';
 import {
   deselectAllFacetValues,
   toggleExcludeFacetValue,
@@ -175,7 +176,7 @@ describe('pagination slice', () => {
   );
 
   it('sets the principal pagination on #fetchProductListing.fulfilled', () => {
-    const response = buildFetchProductListingV2Response({
+    const response = buildFetchProductListingResponse({
       pagination,
     });
 
@@ -244,6 +245,10 @@ describe('pagination slice', () => {
 
   describe.each([
     {
+      actionName: '#deselectAllBreadcrumbs',
+      action: deselectAllBreadcrumbs,
+    },
+    {
       actionName: '#deselectAllFacetValues',
       action: deselectAllFacetValues,
     },
@@ -279,8 +284,7 @@ describe('pagination slice', () => {
     it('resets principal pagination', () => {
       state.principal.page = 5;
 
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const finalState = paginationReducer(state, action({} as any));
+      const finalState = paginationReducer(state, action({} as never));
 
       expect(finalState.principal.page).toBe(0);
     });