+++ packages/headless/src/features/facets/facet-set/facet-set-selectors.ts
@@ -1,12 +1,12 @@
 import {
-  ProductListingSection,
+  OldProductListingSection,
   SearchSection,
 } from '../../../state/state-sections';
 import {FacetSection} from '../../../state/state-sections';
 import {AnyFacetResponse} from '../generic/interfaces/generic-facet-response';
 import {FacetResponse, FacetValue} from './interfaces/response';
 
-export type FacetResponseSection = SearchSection | ProductListingSection;
+export type FacetResponseSection = SearchSection | OldProductListingSection;
 
 export const baseFacetResponseSelector = (
   state: Partial<FacetResponseSection>,